build: enforce fatal warnings and clean RP2040 extra diagnostics
Centralize fatal warning enforcement outside resettable CFLAGS, fix the RP2040 kernel diagnostics without group-wide suppressions, and add calibrated compiler-command regression gates.

+++ sys/arch/rp2040/dhara/map.c
@@ -119,7 +119,7 @@ static int trace_path(struct dhara_map *m, dhara_sector_t target,
 		      dhara_error_t *err)
 {
 	uint8_t meta[DHARA_META_SIZE];
-	int depth = 0;
+	unsigned int depth = 0;
 	dhara_page_t p = dhara_journal_root(&m->journal);
 
 	if (new_meta)
@@ -447,7 +447,7 @@ static int try_delete(struct dhara_map *m, dhara_sector_t s,
 	meta_set_id(meta, meta_get_id(alt_meta));
 
 	meta_set_alt(meta, level, DHARA_PAGE_NONE);
-	for (i = level + 1; i < DHARA_RADIX_DEPTH; i++)
+	for (i = level + 1; (unsigned int)i < DHARA_RADIX_DEPTH; i++)
 		meta_set_alt(meta, i, meta_get_alt(alt_meta, i));
 
 	meta_set_alt(meta, level, DHARA_PAGE_NONE);
